Physics source tree update
The physics system is pending ANOTHER rewrite.

See diagram for information on how - we will likely use bevy, which
would entail a full source tree rewrite.

Maybe the spring system code can be reused, but it is likely not worth
it.

All the cross-references (dotted gray lines) between different
components in the source tree would be handled via slotmap keys. Some
of the references are multi-references, and thus use SmallVec of said
keys.